IMX-2200 CL280517 CL280598 Change GBM_MAX_BUFFER to 3 from 4 for avoid memory "wasting"

+++ driver/driver/khronos/libVulkan11/src/chip/gc_halti5_cmdbuf.c
@@ -10998,7 +10998,7 @@ static VkResult halti5_helper_setDescSetStorage(
                         uint32_t offset = (uint32_t)resInfo->u.bufferInfo.offset + dynamicOffsets[(*dynamicOffsetIndex) + arrayIdx];
                         __vkBuffer *buf = resInfo->u.bufferInfo.buffer;
                         physical = buf->memory->devAddr;
-                        offset += buf->memOffset;
+                        offset += (uint32_t)buf->memOffset;
                         physical += offset;
                         data[dataCount++] = physical;
                         if (devCtx->enabledFeatures.robustBufferAccess)
